% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Title}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{default}

\item{pkgname}{default}
}
\description{
add resource paths for images, javascript and css files
}
\author{
Dorlan Francis
}
