% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onUnload}
\alias{.onUnload}
\title{Title}
\usage{
.onUnload(libname, pkgname)
}
\arguments{
\item{libname}{default}

\item{pkgname}{default}
}
\description{
removes resource paths that were load by the onLoad function
}
