% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{headerLinks}
\alias{headerLinks}
\title{Header Links}
\usage{
headerLinks()
}
\value{
HTML Header Tag with link tags for Favicon, CSS, Javascript
}
\description{
This utility function is where all html links made in the \verb{<head>}  tag will be placed.
}
\author{
Dorlan Francis
}
