% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_About_Page.R
\name{mod_About_Page_UI}
\alias{mod_About_Page_UI}
\title{About Page}
\usage{
mod_About_Page_UI(id)
}
\arguments{
\item{id}{Modual ID character/string}
}
\value{
HTML Tags
}
\description{
This module returns the about page
}
\author{
Dorlan Francis
}
