% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Footer.R
\name{mod_Footer_Server}
\alias{mod_Footer_Server}
\title{App Footer Module Server Code}
\usage{
mod_Footer_Server(id)
}
\arguments{
\item{id}{Unique id use with ns()}
}
\description{
App Footer Module Server Code
}
