% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Footer.R
\name{mod_Footer_UI}
\alias{mod_Footer_UI}
\title{App Footer Module UI}
\usage{
mod_Footer_UI(id)
}
\arguments{
\item{id}{Unique id use with ns()}
}
\value{
HTML div wrapper for footer elements
}
\description{
App Footer Module UI
}
