% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Leaflet_Map.R
\name{mod_Leaflet_Map_Server}
\alias{mod_Leaflet_Map_Server}
\title{Leaflet Map Module Server}
\usage{
mod_Leaflet_Map_Server(id)
}
\arguments{
\item{id}{id is use with ns('') to uniquely identify module specific ids}
}
\value{
Search Area
}
\description{
Leaflet Map Module Server
}
