% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Leaflet_Map.R
\name{mod_Leaflet_Map_UI}
\alias{mod_Leaflet_Map_UI}
\title{Leaflet Map Module UI}
\usage{
mod_Leaflet_Map_UI(id, width = "100\%", height = "85vh")
}
\arguments{
\item{id}{is use with ns('') to uniquely identify module specific ids}

\item{width}{of map output default is 100\% of container}

\item{height}{of map default is 85\% of view port height}
}
\value{
Leaflet Map rendered output
}
\description{
Leaflet Map Module UI
}
