% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Page_Model_Info.R
\name{mod_Page_Model_Info_UI}
\alias{mod_Page_Model_Info_UI}
\title{About Model Search Info Page}
\usage{
mod_Page_Model_Info_UI(id)
}
\arguments{
\item{id}{Modual ID character/string}
}
\value{
HTML Tags
}
\description{
This module returns the Primary Research info Page#'
}
\author{
Dorlan Francis
}
