% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Primary_Research_Info_Page.R
\name{mod_Primary_Research_Info_Page_UI}
\alias{mod_Primary_Research_Info_Page_UI}
\title{About Primary Research Info Page}
\usage{
mod_Primary_Research_Info_Page_UI(id)
}
\arguments{
\item{id}{Modual ID character/string}
}
\value{
HTML Tags
}
\description{
This module returns the Primary Research info Page
}
