% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{RiCHT_Theme}
\alias{RiCHT_Theme}
\title{Rich Theme (this is about setting colours of the app)}
\usage{
RiCHT_Theme()
}
\value{
fresh theme
}
\description{
Rich Theme (this is about setting colours of the app)
}
