% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny App}
\usage{
run_app(host = NULL, port = NULL)
}
\arguments{
\item{host}{character specify host name}

\item{port}{numeric specify port number}
}
\description{
Run Shiny App
}
\author{
Dorlan Francis
}
