% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Title_Logo.R
\name{mod_Title_Logo_UI}
\alias{mod_Title_Logo_UI}
\title{Title & Logo for Atmel Temple Dashboard}
\usage{
mod_Title_Logo_UI(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Title & Logo for Atmel Temple Dashboard
}
